+++ Pangle/src/main/java/com/mopub/nativeads/PangleAdViewBinder.java
@@ -2,7 +2,6 @@ package com.mopub.nativeads;
 
 import androidx.annotation.NonNull;
 
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
